[MOVE/#11] Friend 상위폴더 이동
